+++ R/report_intercept.R
@@ -76,7 +76,7 @@ print.report_intercept <- function(x, ...) {
 #' @keywords internal
 .find_intercept <- function(model) {
   # Intercept-only
-  if (insight::is_nullmodel(model)) {
+  if (suppressWarnings(insight::is_nullmodel(model))) {
     return("")
   }