+++ tests/testthat/test-format_model.R
@@ -7,7 +7,7 @@ test_that("format_model", {
 
 
 test_that("format_model", {
-  skip_if_not_or_load_if_installed("lme4")
+  skip_if_not_installed("lme4")
   expect_identical(format_model(lme4::lmer(wt ~ cyl + (1 | gear), data = mtcars)), "linear mixed model")
   expect_identical(
     format_model(lme4::glmer(vs ~ cyl + (1 | gear), data = mtcars, family = "binomial")),
@@ -23,18 +23,26 @@ test_that("format_model", {
 
 
 test_that("format_model", {
-  skip_if_not_or_load_if_installed("rstanarm")
+  skip_if_not_installed("rstanarm")
   expect_identical(
-    format_model(suppressWarnings(rstanarm::stan_glm(mpg ~ wt, data = mtcars, refresh = 0, iter = 50))),
+    format_model(
+      suppressWarnings(
+        rstanarm::stan_glm(mpg ~ wt, data = mtcars, refresh = 0, iter = 50)
+      )
+    ),
     "Bayesian linear model"
   )
   expect_identical(
-    format_model(suppressWarnings(rstanarm::stan_glm(vs ~ wt,
-      data = mtcars,
-      family = "binomial",
-      refresh = 0,
-      iter = 50
-    ))),
+    format_model(
+      suppressWarnings(
+        rstanarm::stan_glm(vs ~ wt,
+          data = mtcars,
+          family = "binomial",
+          refresh = 0,
+          iter = 50
+        )
+      )
+    ),
     "Bayesian logistic model"
   )
 })